+++ typechain/factories/IDOVOption__factory.ts
@@ -291,9 +291,9 @@ const _abi = [
                 type: "uint128",
               },
               {
-                internalType: "uint16",
+                internalType: "uint104",
                 name: "premiumRate",
-                type: "uint16",
+                type: "uint104",
               },
               {
                 internalType: "address",
@@ -323,9 +323,9 @@ const _abi = [
                 type: "uint128",
               },
               {
-                internalType: "uint16",
+                internalType: "uint104",
                 name: "premiumRate",
-                type: "uint16",
+                type: "uint104",
               },
               {
                 internalType: "address",
